requirements.txt
Flask==3.0.0
selenium==4.15.2
beautifulsoup4==4.12.2
reportlab==4.0.7
python-dotenv==1.0.0
Pillow==12.0.0
webdriver-manager==4.0.1
requests==2.31.0
authlib==1.3.0
Flask-Session==0.5.0
google-generativeai==0.3.2
playwright==1.40.0
